+++ model_training.py
@@ -1,0 +1,30 @@
+import joblib
+from sklearn.model_selection import train_test_split
+from sklearn.ensemble import GradientBoostingClassifier
+from sklearn.metrics import accuracy_score, roc_auc_score
+from data_pipeline import load_and_preprocess_data
+
+# Load the preprocessed data
+X_scaled, y = load_and_preprocess_data()
+
+# Train-test split
+X_train, X_test, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)
+
+# Train the Gradient Boosting classifier
+clf = GradientBoostingClassifier(random_state=42)
+clf.fit(X_train, y_train)
+
+# Make predictions on the test set
+y_pred = clf.predict(X_test)
+y_prob = clf.predict_proba(X_test)[:, 1]  
+
+# Evaluate the model
+accuracy = accuracy_score(y_test, y_pred)
+roc_auc = roc_auc_score(y_test, y_prob)
+
+print(f"Accuracy: {accuracy:.4f}")
+print(f"ROC AUC: {roc_auc:.4f}")
+
+joblib.dump(clf, './models/diabetes_model.pkl')
+
+print("Model saved as 'diabetes_model.pkl'")
